I need to make smaller commits
But not today.

Clean up on_play_pause() calls. Just pass a clone of the episode instead of a handfull of fields. This also makes the function logic a little cleaner.

Change time format funcs to take i32 because the db stores listen duration as a i32 anyway.

Change podcastid to i32. I'm not entirely sure why I did this, but there was a good reason. 2 billion ids should be more than enough.

Replace more {Page}Episode types with Episode. This should take care of all of them.

Fix deserializing Episodeartwork to correct field

Rename VEpisode to EpisodeListItem

Clean up some commented-out code and superfluous comments

Consolidate some {Page}VirtualList structs into a single VirtualList. This reduces some redundant args, but requires a new struct DragCallbacks to enable dragging on a VirtualList.

+++ web/src/components/click_events.rs
@@ -15,7 +15,7 @@ pub fn create_on_title_click(
     server_name: String,
     api_key: Option<Option<String>>,
     history: &BrowserHistory,
-    podcast_index_id: i64,
+    podcast_index_id: i32,
     podcast_title: String,
     podcast_url: String,
     podcast_description: String,
